rename DBconnect
