+++ c_api_dps/Dps/Debug/Dps.Build.CppClean.log
@@ -1,6 +1,6 @@
 c:\users\mcdfreedom\kemyond\c_api_dps\dps\debug\vc120.pdb
 c:\users\mcdfreedom\kemyond\c_api_dps\dps\debug\vc120.idb
-c:\users\mcdfreedom\kemyond\c_api_dps\dps\debug\tools.obj
+c:\users\mcdfreedom\kemyond\c_api_dps\dps\debug\dbconnect.obj
 c:\users\mcdfreedom\kemyond\c_api_dps\dps\debug\dps.tlog\cl.command.1.tlog
 c:\users\mcdfreedom\kemyond\c_api_dps\dps\debug\dps.tlog\cl.read.1.tlog
 c:\users\mcdfreedom\kemyond\c_api_dps\dps\debug\dps.tlog\cl.write.1.tlog
