+++ c_api_dps/Dps/Debug/Dps.log
@@ -1,4 +1,4 @@
-﻿生成启动时间为 2018/4/19 16:21:40。
+﻿生成启动时间为 2018/4/19 16:50:14。
      1>项目“C:\Users\McdFreedom\kemyond\c_api_dps\Dps\Dps.vcxproj”在节点 2 上(Build 个目标)。
      1>ClCompile:
          D:\software\Microsoft Visual Studio 12.0\VC\bin\CL.exe /c /IC:\Users\McdFreedom\kemyond\c_api_dps\third\inc /ZI /nologo /W3 /WX- /Od /Oy- /D WIN32 /D _DEBUG /D _CONSOLE /D _LIB /D _UNICODE /D UNICODE /Gm /EHsc /RTC1 /MDd /GS /fp:precise /Zc:wchar_t /Zc:forScope /Fo"Debug\\" /Fd"Debug\vc120.pdb" /Gd /TP /analyze- /errorReport:prompt main.cpp
@@ -6,10 +6,10 @@
      1>c:\users\mcdfreedom\kemyond\c_api_dps\dps\main.cpp(75): warning C4018: “<”: 有符号/无符号不匹配
        Link:
          D:\software\Microsoft Visual Studio 12.0\VC\bin\link.exe /ERRORREPORT:PROMPT /OUT:"C:\Users\McdFreedom\kemyond\c_api_dps\Debug\Dps.exe" /INCREMENTAL /NOLOGO /LIBPATH:C:\Users\McdFreedom\kemyond\c_api_dps\third\dblib libmysql.lib kernel32.lib user32.lib gdi32.lib winspool.lib comdlg32.lib advapi32.lib shell32.lib ole32.lib oleaut32.lib uuid.lib odbc32.lib odbccp32.lib kernel32.lib user32.lib gdi32.lib winspool.lib comdlg32.lib advapi32.lib shell32.lib ole32.lib oleaut32.lib uuid.lib odbc32.lib odbccp32.lib /MANIFEST /MANIFESTUAC:"level='asInvoker' uiAccess='false'" /manifest:embed /DEBUG /PDB:"C:\Users\McdFreedom\kemyond\c_api_dps\Debug\Dps.pdb" /SUBSYSTEM:CONSOLE /TLBID:1 /DYNAMICBASE /NXCOMPAT /IMPLIB:"C:\Users\McdFreedom\kemyond\c_api_dps\Debug\Dps.lib" /MACHINE:X86 Debug\main.obj
-         Debug\Tools.obj
+         Debug\DBConnect.obj
          Dps.vcxproj -> C:\Users\McdFreedom\kemyond\c_api_dps\Debug\Dps.exe
      1>已完成生成项目“C:\Users\McdFreedom\kemyond\c_api_dps\Dps\Dps.vcxproj”(Build 个目标)的操作。
 
 生成成功。
 
-已用时间 00:00:01.82
+已用时间 00:00:01.83
